MERN-Training-Test-Engine/backend/requirements.txt
django>=4.0
djangorestframework>=3.0
django-cors-headers